- Fix wxString::Format() modifier, invalid when using 64-bit.
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@12997 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/toolsmanager.cpp
@@ -264,7 +264,8 @@ void ToolsManager::LoadTools()
 
         AddTool(&tool, false);
     }
-    Manager::Get()->GetLogManager()->Log(wxString::Format(_("Configured %d tools"), m_Tools.GetCount()));
+
+    Manager::Get()->GetLogManager()->Log(wxString::Format(_("Configured %lu tools"), (unsigned long)m_Tools.GetCount()));
 }
 
 void ToolsManager::SaveTools()
